Run ASan and UBsan and fix issues
Finally I found out how to run the sanitizers from clang.

This commit fixes a ton of memory leaks, UB and memory errors. The
alignment sanitizer is still broken and thus disabled for now.

+++ rir/src/compiler/native/builtins.cpp
@@ -286,7 +286,7 @@ static SEXP callImplCached(ArglistOrder::CallId callId, rir::Code* c,
                            unsigned long available, Immediate cache) {
     auto ctx = globalContext();
     CallContext call(callId, c, callee, nargs, ast,
-                     ostack_cell_at(ctx, nargs - 1), env, R_NilValue,
+                     ostack_cell_at(ctx, (long)nargs - 1), env, R_NilValue,
                      Context(available), ctx);
 
     SLOWASSERT(env == symbol::delayedEnv || TYPEOF(env) == ENVSXP ||
